Refactor imports for @hedvig-ui/Dropdown
