+++ shared/hedvig-ui/Dropdown/dropdown.stories.tsx
@@ -1,4 +1,4 @@
-import { EnumDropdown } from 'hedvig-ui/dropdown'
+import { EnumDropdown } from '@hedvig-ui'
 import React from 'react'
 
 export default {
