+++ src/components/claims/claim-details/components/ClaimInformation.tsx
@@ -7,7 +7,13 @@ import {
   useClaimPageQuery,
 } from 'api/generated/graphql'
 
-import { CardContent, CardsWrapper, DangerCard } from '@hedvig-ui'
+import {
+  CardContent,
+  CardsWrapper,
+  DangerCard,
+  Dropdown,
+  EnumDropdown,
+} from '@hedvig-ui'
 import { PaperTitle } from 'components/claims/claim-details/components/claim-items/PaperTitle'
 import { ContractDropdown } from 'components/claims/claim-details/components/ContractDropdown'
 import { format, parseISO } from 'date-fns'
@@ -23,7 +29,6 @@ import {
   updateClaimStateOptions,
   useUpdateClaimState,
 } from 'graphql/use-update-claim-state'
-import { Dropdown, EnumDropdown } from 'hedvig-ui/dropdown'
 import { InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { Loadable } from 'hedvig-ui/loadable'
 import { Label, Paragraph } from 'hedvig-ui/typography'
