+++ src/components/member/tabs/quote-tab/index.tsx
@@ -1,3 +1,4 @@
+import { getTextFromEnumValue } from '@hedvig-ui'
 import { Quote } from 'api/generated/graphql'
 import { useQuotes } from 'graphql/use-get-quotes'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
@@ -5,7 +6,6 @@ import {
   LoadingMessage,
   StandaloneMessage,
 } from 'hedvig-ui/animations/standalone-message'
-import { getTextFromEnumValue } from 'hedvig-ui/dropdown'
 import React from 'react'
 import { Tab } from 'semantic-ui-react'
 import { ContractType, Market, QuoteProductType } from 'types/enums'
